server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
rembg==2.0.59
Pillow==10.4.0
